requirements.txt
# Tumblr & HTTP
pytumblr>=0.1.0
requests>=2.28.0
requests-oauthlib>=1.3.0

# Web app
flask>=3.0.0
flask-cors>=4.0.0

# Data & scheduling
python-dateutil>=2.8.0

# Env
python-dotenv>=1.0.0
